Adjust to the new mail API
The new API returns an HTML table instead of a text which we now parse using
Beautifulsoup4.

The use of json.loads() instead of response.json() is due to the fact that
response.json() raises an exception when it receives a mime type which it does
not expect.

I also fixed some pylint errors and sorted the imports using isort.